[feat] fixed account creation bug and enabled facebook and google login
