+++ dev/reducers/user_reducer.js
@@ -7,13 +7,6 @@ export default function(state = initialState, action) {
       return Object.assign({},state,{
           userProfile: action.payload
         });
-    case 'USER_CREATED':
-      console.log('user created reducer')
-      return Object.assign({}, state, {
-        isAuthenticated: true,
-        profile: action.profile,
-        error: ''
-      });
     default:
       return state;
   }
